Fully parameterize the tsem_inode_create handler.
The following structure was added in order to support the
retentation and translation of the arguments to the handler for
the inode_create LSM hook:

tsem_inode_create_args

The get_inode_create() function was added to the event.c file in
order to translate the input arguments to the retention
arguments.  Since the mode value can be retained in form it was
placed in the structure and not the in/out structures.

This requires that only the 'out' structure be zeroed since
zeroing of the entire structure, as was done previously wipes out
the mode argument.  This suggests the opportunity to reform the
other argument filling functions since there is no need to clear
the entire structure.

The show_inode_create() function was added to the fs.c file to
encapulate the generation of the event description for the
handler.

Since the mode value is a 16 bit integer, and will be used by
other inode_* arguments the following function was added to the
map.c file:

add_16()

Which as it's name implies, extends the security event
coefficient with the contents of a 16 bit integer value.