Convert tsem_inode_rmdir to full parameterization.
The tsem_inode_create_args structure was once again re-used to
fully parameterize this event handler.  This suggests that as the
work completes an investigation into a commonly named structure
may be advisable.

While the argument structure could be re-used an alternate
mapping clause was needed since there was no mode value to be
extended into the generation of the coefficient.

The following new function was added for displying the CELL value
of the event for the same reason:

fs.c:show_inode_remove()