Fully parameterize the tsem_inode_rename handler.
Due to the number of arguments entertained by the handler for the
inode_rename LSM call the tsem_inode_rename_args structure was
introduced to carry the input and retention arguments for the
handler.

A new switch clause was added to the get_cell_mapping() function
in order to create the security coefficient based on the event
parameters.

The event.c:get_inode_rename() function was implemented in order
to convert the input arguments to their retained version.

The fs.c:show_inode_rename() function was implemented in order to
encapsulate the generation of the security event description.