Parameterize the tsem_task_setpgid handler.
The fs.c:show_task_setpgid() function was added to generate the
security state description for the event.

A new clause was added to the switch statement in the
map.c:get_cell_mapping() to compute the security state
coefficient for the event.

Special handling was implemented in the event handler itself to
address the issue of the pgid argument passed to the handler.
This argument specifies the process group that the target process
is to be placed into.

If this value is zero the process group of the current process is
isued.  If the value is non-zero the argument specifies the PID
of the process whose process group is to be used.

Since process ID's are non-deterministic TSEM uses the task
identity of a process to identify the process.  If the pgid value
is non-zero the task structure of the process to be used is
looked up and the task identity of that process used as the
'source' argument.

The TSEM handler function was not using the standardized
infrastructure for reporting on a non-trusted process.  This
commit standardizes the handling of an untrusted event.