Parameterize the unix_stream_connect and unix_send handlers.
The conversion of the TSEM handlers for the LSM hooks that handle
AF_UNIX connections was done in tandem since their implementation
is largely identical, with the events being differentiated by the
identity of the handler.

This conversion involves the introduction of general
infrastructure in the form of a handler for the following LSM
hook:

socket_post_create

This LSM hook is invoked after socket creation in order to allow
security attributes to be added to a socket after its creation.

The purpose of the TSEM implementation of this handler is to
capture the task identity of the process that is creating the
socket.  The task identity is captured into a buffer that was
added to the tsem_socket_create_args structure.  The notion of
the ownership of the socket will be propagated to other socket
hooks at a later time.

The task identity is incorporated into the inode that represents
the socket so all other inode related creation events will
inherit this functionality.  For the file based inode creation
events this functionality will be leveraged to address the issue
of files that are created in a security modeling namespace and
thus have no fixed file identity with respect to the contents of
the file.

The following globally visible structure was also added to
support this work:

tsem_unix_socket_args

Which allows the incoming argument pointers to be transferred to
the tsem_socket_create_args structure that are embedded in the
out portion of the tsem_unix_socket_args.

A common clause was added to the switch statement in the
map.c:get_cell_mapping() function to generate the security state
coefficient for each of these event handlers.

These were accompanied by the addition of the
map.c:add_unix_socket() function which encapsulates the
generation of the extension stream for a single instance of a
socket.

The fs.c:show_unix_socket() function was introduced to output the
security event description for both of these handlers.