Fully parameterize the tsem_socket_socketpair handler.
The parameterization of this security handler suggested further
generalization of the infrastructure for handling security events
that involve multiple sockets.  In fact the parameterization of
this event largely re-uses infrastructure that was implemented
for the tsem_unix_stream_connect and tsem_unix_may_send handlers.

Rather than use the 'sock' and 'other' members tsem_socket_args
structure the members were named to 'socka' and 'sockb'.  This is
the convention that is used in the socket_socketpair LSM hook and
is a more general approach.

The fs.c:show_unix_socket() function was renamed to
show_socket_pair().

In a similar approach the get_unix_socket() function was renamed
to get_socket_pair().