Fully parameterize the tsem_settime handler.
The following structure was introduced into the tsem.h file to
hold the arguments to the security_settime hook:

tsem_time_args

Since these were contants values there is no need to conduct the
arguments into a retention structure.

The fs.c:show_settime() function was introduced in order to
display the security event description for this event.

A clause was added to the switch statement in the
map.c:get_cell_mapping() function in order to generate the
security state coefficient for the event.

Testing indicates that the struct timezone pointer may be a null
value.  A condition check was implemented to skip collection of
the arguments for the settime function in this case.

Out of an abundance of caution a condition was also added to the
struct timespec64 pointer passed to the handler.