Generate security state coeficient for inode_listxattr.
Inspection of code secondary to the parameterization of the
sb_statfs handler found that the parameterization of the
inode_listxattr handler had neglected to include a clause in the
map.c:get_cell_mapping() function to generate a security
coefficient for the parameterized event.

Since the inode_listxattr handler takes as an argument a single
dentry, the same as the sb_statfs handler, include with the
TSEM_SB_STATFS case a TSEM_INODE_LISTXATTR case.