Full parameterize the tsem_key_alloc handler.
The following structure was added to the tsem.h global header
file to hold arguments passed to and retained to the security
handlers for key relevant events:

tsem_key_args

A separate clause was added to the switch statement in the
map.c:get_cell_mapping() function to compute the security
coefficient for the key allocation event.

The event.c:get_key_alloc() function was introduced to transfer
the credentials passed to the tsem_key_alloc security handler
into their retention variants.

The following function was introduced to display the security
event description for an instance of this handler:

fs.c:show_key_alloc()