Fully parameterize the tsem_sb_remount handler.
This completes the parameterization of the security handlers that
involve a filesystem superblock.

In order to parameterize this event a pointer to a super_block
structure was added to the 'in' portion of the tsem_sb_args
structure.

The event.c:get_sb_remount() function was implemented to convert
the incoming arguments for the sb_remount handler into their
retained values.

A clause was added to the switch statment in the
event.c:free_cell() function to free the replicated value of the
filesystem type and the path description.

A clause was added to the map.c:get_cell_mapping() function to
compute the security state coefficient for the event.

The fs.c:show_sb_remount() function was implemented to generate
the security description for the event.