Handle NULL arguments for device name and filesystem type.
Testing on Debian indicated segmentation faults secondary to the
userspace invoking the mount system call and passing a NULL
pointer in as either the device name or the filesystem type.

This was addressed by testing for these NULL values and not
conducting the parameter replication for these arguments in the
event.c:get_sb_mount() function.

This left NULL pointers for the retention arguments.  Check for
the presence of these NULL pointers in the
map.c:get_cell_mapping() function code so that coefficient
generation does not attempt to extend the coefficient state with
these arguments if they were presented as NULL values.

Similarily in the fs.c:show_sb_mount() function, check for the
presence of a NULL pointer and refrain from displaying the values
if they were not presented.